prace wre

+++ asserts.h
@@ -7,11 +7,11 @@ void assertMatrixRGB(const cv::Mat &matrix) {
 }
 
 void assertMatrixNormalizedRGB(const cv::Mat &matrix) {
-	CV_Assert(matrix.type() == CV_8UC3);
+	CV_Assert(matrix.type() == CV_32FC3);
 }
 
 void assertMatrixGrayScale(const cv::Mat &matrix) {
-	CV_Assert(matrix.type() == CV_8UC1);
+	CV_Assert(matrix.type() == CV_32FC1);
 }
